Replace insecure Math.random() with crypto.getRandomValues()
`Math.random()` is not cryptographically secure, and as such is not suitable for
generating passwords. As such, it is replaced with `crypto.getRandomValues()`.

`Math.floor(Math.random * length)` is also not uniform if `length` is not a
multiple of a power of 2. As such, a `secureRandom(count)` function was built
which uses modulo with rejection sampling to ensure every character or word when
building the password or passphrase is uniform. Every instance of the
multiply-and-floor method was replaced with this function.

These changes bring the password and passphrase generators up to standard with
modern cryptographic best practices.